refactor(coordinator): split worker manager responsibilities
